Introduce command pattern and Null command
Each command issued from the user will cause a command object to be
created. This object will be responsible for validating the command and
for executing the commands actions.

All input will result in a new command object, defaulting to the Null
command.

Fix spec_helper file loading within sub-directories.