Add CommandInterpreter class
This is used to handle lines on input and create the corresponding
command object.

Needs to pass in any arguments to the class initializer as a Array that
will be mapped to a hash according to the command class's argument
requirements.

The command also needs to have knowledge of the Simulation for it to be
able to interact with the objects within.